+++ packages/transducers/src/xform/drop-nth.ts
@@ -1,3 +1,4 @@
+import { isIterable } from "@thi.ng/checks";
 import { iterator1 } from "../iterator";
 import { throttle } from "./throttle";
 import type { Transducer } from "../api";
@@ -5,7 +6,7 @@ import type { Transducer } from "../api";
 export function dropNth<T>(n: number): Transducer<T, T>;
 export function dropNth<T>(n: number, src: Iterable<T>): IterableIterator<T>;
 export function dropNth<T>(n: number, src?: Iterable<T>): any {
-    if (src) {
+    if (isIterable(src)) {
         return iterator1(dropNth(n), src);
     }
     n = Math.max(0, n - 1);