+++ packages/transducers/src/xform/filter.ts
@@ -1,3 +1,4 @@
+import { isIterable } from "@thi.ng/checks";
 import { compR } from "../func/compr";
 import { iterator1 } from "../iterator";
 import type { Predicate } from "@thi.ng/api";
@@ -9,7 +10,7 @@ export function filter<T>(
     src: Iterable<T>
 ): IterableIterator<T>;
 export function filter<T>(pred: Predicate<T>, src?: Iterable<T>): any {
-    return src
+    return isIterable(src)
         ? iterator1(filter(pred), src)
         : (rfn: Reducer<any, T>) => {
               const r = rfn[2];